refactor: action btn to its own component
