+++ web/assets/js/react/modules/History/ProofHistory.tsx
@@ -11,9 +11,9 @@ import { ProofSubmission, SubmitProof } from "../../types/aligned";
 import { timeAgo } from "../../utils/date";
 import { computeVerificationDataCommitment } from "../../utils/aligned";
 import { shortenHash } from "../../utils/crypto";
-import { Button } from "../../components";
 import { useCSRFToken } from "../../hooks/useCSRFToken";
 import { useProofSentMessageReader } from "../../hooks/useProofSentMessageReader";
+import { ProofEntryActionBtn } from "./ProofEntry";
 
 const colorBasedOnStatus: { [key in ProofSubmission["status"]]: string } = {
 	submitted: "bg-accent-100/20 text-accent-100",
@@ -46,13 +46,6 @@ const statusText: { [key in ProofSubmission["status"]]: string } = {
 	failed: "Failed",
 };
 
-const actionBtn: { [key in ProofSubmission["status"]]: string } = {
-	claimed: "Share",
-	submitted: "Submit solution",
-	pending: "Bump fee",
-	failed: "None",
-};
-
 type Props = {
 	leaderboard_address: Address;
 	user_address: Address;
@@ -68,9 +61,6 @@ export const ProofHistory = ({
 }: Props) => {
 	const [proofsTableRows, setProofsTableRows] = useState<ColumnBody[]>([]);
 	const { csrfToken } = useCSRFToken();
-	const formRetryRef = useRef<HTMLFormElement>(null);
-	const formSubmittedRef = useRef<HTMLFormElement>(null);
-	const [submitProofMessage, setSubmitProofMessage] = useState("");
 	useProofSentMessageReader();
 
 	const { balance } = useBatcherPaymentService({
@@ -78,13 +68,10 @@ export const ProofHistory = ({
 		userAddress: user_address,
 	});
 
-	const { score, submitSolution } = useLeaderboardContract({
+	const { score } = useLeaderboardContract({
 		userAddress: user_address,
 		contractAddress: leaderboard_address,
 	});
-
-	const { estimateMaxFeeForBatchOfProofs, signVerificationData } =
-		useAligned();
 
 	useEffect(() => {
 		const historyBalance = document.getElementById("history-balance");
@@ -95,77 +82,6 @@ export const ProofHistory = ({
 		if (historyScore)
 			historyScore.innerText = score.data?.toString() || "...";
 	}, [balance, score]);
-
-	useEffect(() => {
-		if (submitSolution.receipt.isSuccess) {
-			window.setTimeout(() => {
-				formSubmittedRef.current?.submit();
-			}, 1000);
-		}
-	}, [submitSolution.receipt]);
-
-	const handleBtnClick = (proof: ProofSubmission) => async () => {
-		if (proof.status === "failed") {
-			// nothing to do
-			return;
-		}
-
-		if (proof.status === "claimed") {
-			const text = encodeURIComponent(
-				"🟩 I just claimed my proof on zk-arcade!\n\n"
-			);
-			const url = encodeURIComponent("Try: https://zkarcade.com\n\n");
-			const hashtags = `\naligned,proof,${proof.verificationData.verificationData.provingSystem}`;
-			const twitterShareUrl = `https://twitter.com/intent/tweet?text=${text}&url=${url}&hashtags=${hashtags}`;
-
-			window.open(twitterShareUrl, "_blank");
-
-			return;
-		}
-
-		if (proof.status === "submitted") {
-			if (!proof.batchData) {
-				alert("Batch data not available for this proof");
-				return;
-			}
-
-			await submitSolution.submitBeastSolution(
-				proof.verificationData.verificationData,
-				proof.batchData
-			);
-			return;
-		}
-
-		if (proof.status === "pending") {
-			const maxFee = await estimateMaxFeeForBatchOfProofs(16);
-			if (!maxFee) {
-				alert("Could not estimate max fee");
-				return;
-			}
-
-			proof.verificationData.maxFee = toHex(maxFee, { size: 32 });
-
-			const { r, s, v } = await signVerificationData(
-				proof.verificationData,
-				payment_service_address
-			);
-
-			const submitProofMessage: SubmitProof = {
-				verificationData: proof.verificationData,
-				signature: {
-					r,
-					s,
-					v: Number(v),
-				},
-			};
-
-			setSubmitProofMessage(JSON.stringify(submitProofMessage));
-
-			window.setTimeout(() => {
-				formRetryRef.current?.submit();
-			}, 1000);
-		}
-	};
 
 	useEffect(() => {
 		const rows: ColumnBody[] = proofs.map(proof => ({
@@ -219,69 +135,17 @@ export const ProofHistory = ({
 				<TableBodyItem
 					text={proof.verificationData.verificationData.provingSystem}
 				/>,
-				<td>
-					<Button
-						variant="contrast"
-						className="text-nowrap text-sm w-full"
-						disabled={proof.status === "failed"}
-						style={{
-							paddingLeft: 0,
-							paddingRight: 0,
-						}}
-						onClick={handleBtnClick(proof)}
-					>
-						{actionBtn[proof.status]}
-					</Button>
-
-					{proof.status === "pending" && (
-						<form
-							ref={formRetryRef}
-							action="/proof/status/retry"
-							method="post"
-							className="hidden"
-						>
-							<input
-								type="hidden"
-								name="_csrf_token"
-								value={csrfToken}
-							/>
-							<input
-								type="hidden"
-								name="submit_proof_message"
-								value={submitProofMessage}
-							/>
-							<input
-								type="hidden"
-								name="proof_id"
-								value={proof.id}
-							/>
-						</form>
-					)}
-					{proof.status == "submitted" && (
-						<form
-							className="hidden"
-							ref={formSubmittedRef}
-							action="/proof/status/submitted"
-							method="POST"
-						>
-							<input
-								type="hidden"
-								name="_csrf_token"
-								value={csrfToken}
-							/>
-							<input
-								type="hidden"
-								name="proof_id"
-								value={proof.id}
-							/>
-						</form>
-					)}
-				</td>,
+				<ProofEntryActionBtn
+					leaderboard_address={leaderboard_address}
+					payment_service_address={payment_service_address}
+					user_address={user_address}
+					proof={proof}
+				/>,
 			],
 		}));
 
 		setProofsTableRows(rows);
-	}, [csrfToken]);
+	}, []);
 
 	return (
 		<div className="overflow-auto" style={{ maxHeight: 500 }}>
